Texture updates, aluminum changes, music box added
