+++ Assets/Tiles/AstralMeteor/AluminumBarItem.png
