+++ Assets/Tiles/AstralMeteor/AluminumOreItem.png
