+++ Content/Items/AstralMeteor/Weapons.AluminumCarbide.cs
@@ -1,5 +1,6 @@
 ﻿using Microsoft.Xna.Framework;
 using Microsoft.Xna.Framework.Graphics;
+using StarlightRiver.Content.Tiles.AstralMeteor;
 using StarlightRiver.Core;
 using System;
 using Terraria;
@@ -78,7 +79,7 @@ namespace StarlightRiver.Content.Items.AstralMeteor
         public override void AddRecipes()
         {
             ModRecipe recipe = new ModRecipe(mod);
-            recipe.AddIngredient(ItemType<AluminumBar>(), 20);
+            recipe.AddIngredient(ItemType<AluminumBarItem>(), 20);
             recipe.AddTile(TileID.Anvils);
             recipe.SetResult(this);
             recipe.AddRecipe();
