+++ Content/Tiles/AstralMeteor/AluminumBricks.cs
@@ -8,37 +8,36 @@ using static Terraria.ModLoader.ModContent;
 
 namespace StarlightRiver.Content.Tiles.AstralMeteor
 {
-	public class AluminumOre : ModTile
+	public class AluminumBrick : ModTile
     {
         public override bool Autoload(ref string name, ref string texture) {
             texture = AssetDirectory.AluminumTile + name;
             return base.Autoload(ref name, ref texture); }
 
         public override void SetDefaults() =>
-            this.QuickSet(0, DustType<Dusts.Electric>(), SoundID.Tink, new Color(156, 172, 177), ItemType<AluminumOreItem>(), true, true, "Aluminum Ore");
+            this.QuickSet(0, DustType<Dusts.Electric>(), SoundID.Tink, new Color(156, 172, 177), ItemType<AluminumBrickItem>(), true, true, "Aluminum Brick");
 
         public override void NearbyEffects(int i, int j, bool closer)
         {
             Lighting.AddLight(new Vector2(i + 0.5f, j + 0.5f) * 16, new Vector3(0.1f, 0.32f, 0.5f) * 0.35f);
 
-            if (Main.rand.Next(40) == 0)
+            if (Main.rand.Next(100) == 0)
                 Dust.NewDustPerfect(new Vector2(i + Main.rand.NextFloat(), j + Main.rand.NextFloat()) * 16, DustType<Dusts.Starlight>(), new Vector2(0, -Main.rand.NextFloat(4, 6)), 0, default, Main.rand.NextFloat(0.4f, 0.7f));
 
-            if (Main.rand.Next(300) == 0)
+            if (Main.rand.Next(500) == 0)
             {
                 Vector2 pos = new Vector2(i + 0.5f, j + 0.5f) * 16;
                 DrawHelper.DrawElectricity(pos, pos + Vector2.One.RotatedByRandom(6.28f) * Main.rand.Next(15, 35), DustType<Dusts.Electric2>(), 0.4f, 9);
             }
         }
 
-        public override void FloorVisuals(Player player) => 
-            player.AddBuff(BuffType<Buffs.Overcharge>(), 120);
+        //public override void FloorVisuals(Player player) => 
+        //    player.AddBuff(BuffType<Buffs.Overcharge>(), 120);
     }
 
-    public class AluminumOreItem : QuickTileItem
+    public class AluminumBrickItem : QuickTileItem
     {
-        public AluminumOreItem() : base("Astral Aluminum", "", "AluminumOre", ItemRarityID.Blue, AssetDirectory.AluminumTile) { }
+        public AluminumBrickItem() : base("Astral Aluminum Brick", "", "AluminumBrick", ItemRarityID.White, AssetDirectory.AluminumTile) { }  //TODO: Fix place type
 
-        public override void SafeSetDefaults() => item.value = Item.sellPrice(0, 0, 2, 0);
     }
 }
